+++ core/protocol/all_services.go
@@ -391,8 +391,12 @@ func newServices(
 	// the end of epoch. Since the engine will reject computations when the program
 	// is marked as ended, it needs to be one of the last service to register on
 	// epoch update, so the computation can happen for this epoch.
-	svcs.referralProgram = referral.NewSnapshottedEngine(svcs.epochService, svcs.broker, svcs.timeService, svcs.marketActivityTracker)
+	svcs.referralProgram = referral.NewSnapshottedEngine(svcs.broker, svcs.timeService, svcs.marketActivityTracker)
 	svcs.snapshotEngine.AddProviders(svcs.referralProgram)
+	// The referral program engine must be notified of the epoch change *after* the
+	// market activity tracker, as it relies on computation that must happen, at
+	// the end of the epoch, in market activity tracker.
+	svcs.epochService.NotifyOnEpoch(svcs.referralProgram.OnEpoch, svcs.referralProgram.OnEpochRestore)
 
 	// setup config reloads for all engines / services /etc
 	svcs.registerConfigWatchers()