+++ CHANGELOG.md
@@ -76,6 +76,7 @@
 - [9147](https://github.com/vegaprotocol/vega/issues/9147) - Add reward multiplier to vesting engine
 - [9234](https://github.com/vegaprotocol/vega/issues/9234) - Add support for transfers out of the vested account
 - [9199](https://github.com/vegaprotocol/vega/issues/9199) - Consider running notional volume to determine referrers and referees tier
+- [9214](https://github.com/vegaprotocol/vega/issues/9214) - Add staking tier on referral program
 
 ### 🐛 Fixes